+++ lib/pubnub.dart
@@ -2,10 +2,7 @@
 /// PubNub Data Streaming Network in a fast and easy manner.
 library pubnub;
 
-export './src/core/keyset.dart' show Keyset;
-export './src/core/uuid.dart' show UUID;
-export './src/core/timetoken.dart' show Timetoken, TimetokenDateTimeExtentions;
-export './src/core/exceptions.dart';
+export './src/core/core.dart';
 export './src/dx/_utils/ensure.dart' show InvariantException;
 
 export './src/dx/channel/channel.dart' show Channel;
@@ -16,8 +13,11 @@ export './src/dx/_endpoints/presence.dart' show HeartbeatResult, LeaveResult;
 export './src/dx/_endpoints/signal.dart' show SignalResult;
 
 export './src/dx/subscribe/subscription.dart' show Subscription;
+export './src/dx/subscribe/envelope.dart'
+    show Envelope, PresenceAction, PresenceEvent;
 export './src/dx/channel/channel_history.dart'
     show PaginatedChannelHistory, ChannelHistory;
+export './src/dx/channel/message.dart' show Message;
 
 export './src/dx/pam/pam.dart'
     show Resource, ResourceType, ResourceTypeExtension, TokenRequest, Token;